created a containers folder to organize src folder
